requirements.txt
astroid==3.3.9
blinker==1.9.0
click==8.1.8
dill==0.4.0
Flask==3.1.0
Flask-WTF==1.2.2
importlib_metadata==8.7.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
platformdirs==4.3.7
psycopg2-binary==2.9.10
pylint==3.3.6
python-dotenv==0.21.1
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.13.2
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.21.0
